Split DatasetKey to own class, which actually kinda makes more sense as a 'model' for now.
